review modif: CnecExtension
